#88 Created bezier_surface.h to use in upcoming BezierSurface object.
